+++ Game/RootLoader.h
@@ -12,7 +12,7 @@ public:
 // 继承的
 public:
 	virtual void OnJoined(size_t id, size_t location) override;
-	virtual void OnUpdate(size_t dt) override;
+	virtual void OnUpdate(float dt) override;
 	virtual void OnKicking() override;
 };
